refactor: rename events file
